[REF] pylint_odoo: Clean code (#402)
* Remove unused options
* Get paths similar to pre-commit way
* Remove deprecated folder
* Move checkers structure
* Merge no_odoo and odoo checkers
* Change name of files and classes
* Merge misc.py and settings.py files
* Add same way than oca-odoo-pre-commit-hooks to look for odoo manifest path
