+++ pytest.ini
@@ -40,8 +40,8 @@ addopts =
 # the src/ directory by removing
 # --pyargs and pylint_odoo from the options here.
 testpaths =
-    src
-    tests
+    ./src/pylint_odoo
+    ./tests
 
 # Idea from: https://til.simonwillison.net/pytest/treat-warnings-as-errors
 filterwarnings =
