+++ src/pylint_odoo/__init__.py
@@ -6,9 +6,8 @@ from .augmentations.main import apply_augmentations
 
 def register(linter):
     """Required method to auto register this checker"""
-    linter.register_checker(checkers.modules_odoo.ModuleChecker(linter))
-    linter.register_checker(checkers.no_modules.NoModuleChecker(linter))
-    linter.register_checker(checkers.format.FormatChecker(linter))
+    linter.register_checker(checkers.odoo_addons.OdooAddons(linter))
+    linter.register_checker(checkers.vim_comment.VimComment(linter))
 
     # register any checking fiddlers
     apply_augmentations(linter)
@@ -17,9 +16,8 @@ def register(linter):
 def get_all_messages():
     """Get all messages of this plugin"""
     all_msgs = {}
-    all_msgs.update(checkers.modules_odoo.ODOO_MSGS)
-    all_msgs.update(checkers.no_modules.ODOO_MSGS)
-    all_msgs.update(checkers.format.ODOO_MSGS)
+    all_msgs.update(checkers.odoo_addons.ODOO_MSGS)
+    all_msgs.update(checkers.vim_comment.ODOO_MSGS)
     return all_msgs
