+++ src/pylint_odoo/checkers/__init__.py
@@ -1,4 +1,2 @@
-# pylint:disable=redefined-builtin
-from . import modules_odoo
-from . import no_modules
-from . import format
+from . import odoo_addons
+from . import vim_comment
