--- src/pylint_odoo/settings.py
@@ -1,28 +1,0 @@
-# pylint plugin global msg number for odoo modules msgs
-BASE_OMODULE_ID = 79
-
-# pylint plugin global msg number for python modules msgs
-BASE_PYMODULE_ID = 80
-
-# pylint plugin global msg number for no modules msgs
-BASE_NOMODULE_ID = 81
-
-# pylint plugin global msg number for format msgs
-BASE_FORMAT_ID = 82
-
-# Manifest files of odoo
-MANIFEST_FILES = [
-    "__manifest__.py",
-    "__odoo__.py",
-    "__openerp__.py",
-    "__terp__.py",
-]
-
-# Message description default
-DESC_DFLT = (
-    "You can review guidelines here: "
-    + "https://github.com/OCA/odoo-community.org/blob/master/website/"
-    + "Contribution/CONTRIBUTING.rst"
-)
-
-CFG_SECTION = "odoolint"
