+++ testing/resources/test_repo/broken_module/models/model_inhe2.py
@@ -25,7 +25,7 @@ class TestModel41(models.Model):
 
 
 class TestModel5(models.Model):
-    _inherit = 'valid.duplicated'  # pylint: disable=R7980
+    _inherit = 'valid.duplicated'  # pylint: disable=consider-merging-classes-inherited
 
 
 class TestModel6(models.Model):
